Bug --> again index out of bounce exception + no buffer read

+++ io.openems.edge.bridge.genibus/src/io/openems/edge/bridge/genibus/api/GenibusTask.java
@@ -4,7 +4,7 @@ package io.openems.edge.bridge.genibus.api;
 public interface GenibusTask {
 
 
-    byte getRequest();
+    int getRequest();
 
     void setResponse(byte data);
