+++ io.openems.edge.heatpump.device/src/io/openems/edge/heatpump/device/task/HeatPumpReadTask.java
@@ -15,7 +15,7 @@ public class HeatPumpReadTask extends AbstractHeatPumpTask {
     }
 
     @Override
-    public byte getRequest() {
+    public int getRequest() {
         return -1;
     }
